requirements.txt
Flask
pandas
requests
faiss-cpu
numpy
python-dotenv
langchain
langchain-openai
